pyproject.toml
[tool.poetry]
name = "lyvego-last-version"
version = "0.1.0"
description = "Lyevgo packages"
authors = ["takitsu21 <dylannbatisse@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = "^1.7.3"
aiomysql = "^0.0.21"
aiohttp = "^3.7.4.post0"
aiocontextvars = "^0.2.2"
discord-py-slash-command = "^2.3.1"
topggpy = "^1.3.0"
aiofiles = "^0.7.0"
python-dotenv = "^0.18.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
